+++ src/main.cpp
@@ -1,1 +1,2 @@
+#include "hooks/CCEGLView.hpp"
 // hehe